데미지클래스추가  이펙트매니저추가 옷입히기  일부 소스정리

+++ PortfolioGame/PortfolioGame/Managers/Skins/skin_manager.h
@@ -28,7 +28,7 @@ public:
 
 private:
 	void LoadBase();
-	void UolSetting(uint16_t skinId, std::map<std::string, SkinInfo*> info, SkinItem* skinitem);
+	void UolSetting(uint32_t skinId, std::map<std::string, SkinInfo*> info, SkinItem* skinitem);
 private:
 	std::map<std::string, SkinInfo*> _bodySkins;
 	std::map<std::string, SkinInfo*> _itemSkins;
